graphql_server/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
grpcio
grpcio-tools
graphene-django
asyncpg